pyproject.toml
[project]
name = "fitness-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "decouple>=0.0.7",
    "django>=6.0a1",
    "django-cors-headers>=4.0.0",
    "django-filter>=25.2",
    "django-jazzmin>=3.0.1",
    "django-rest-framework>=0.1.0",
    "djangorestframework>=3.14.0",
    "drf-spectacular>=0.26.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.5",
    "python-dateutil>=2.8.2",
    "python-decouple==3.8",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=20.0",
    "requests>=2.31.0",
]

[tool.isort]
profile = "hug"
skip = "build,.tox,.venv"
src_paths="isort,test"
